// front-end/src/app/models/currentProspecto.model.ts
export class DatosProspecto {
    id: number;
    nombre: string;
    primer_apellido: string;
    segundo_apellido: string;
    calle: string;
    numero: number;
    colonia: string;
    codigo_postal: number;
    telefono: number;
    rfc: string;
    status_prospecto: string;
    observaciones: string;
}